requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
pydantic
pydantic-settings
email-validator
openpyxl
PyJWT
